Moved meta tags to different component
