+++ components/index.js
@@ -1,2 +1,3 @@
 export { default as Navbar } from "./Navbar";
 export { default as Hero } from "./Hero";
+export { default as MetaTags } from "./MetaTags";
